Fix potential divide by zero error for the DateTime tick provider
When a chart with an AutoAdjustingDateTimeTickProvider on one axis is zoomed in to the point where only one tick is visible, the console gets flooded with divide by zero warnings.

I'm not sure if this will be encountered under usual circumstances; my chart configuration had specific positions for the ticks. Still, it's seems like a good safety measure.